+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 
 ## [Unreleased]
 
+- Update `mise@latest` to 2026.5.0.
+
 - Support `rust`. ([#1779](https://github.com/taiki-e/install-action/pull/1779))
 
   This installs rust using rustup.